backend/requirements.txt
flask
flask-cors
nltk
scikit-learn
gensim
numpy
tqdm
faiss-cpu
